+++ class/scrumtask.class.php
@@ -527,10 +527,10 @@ class ScrumTask extends CommonObject
 
 		if($key == 'qty_planned' ||  $key == 'qty_consumed')
 		{
-			if ( !function_exists('getTileFormatedTime')) {
+			if ( !function_exists('convertFloatHourToHoursMins')) {
 				include_once __DIR__ . "/../lib/scrumproject.lib.php" ;
 			}
-			$out =  getTileFormatedTime($value) ;
+			$out =  convertFloatHourToHoursMins($value,$langs) ;
 		} else{
 			$out = parent::showOutputField($val, $key, $value, $moreparam, $keysuffix, $keyprefix, $morecss);
 		}